member package명 수정
