+++ src/main/java/Spring/Goods_Shop/member/MasterController.java
@@ -1,4 +1,4 @@
-package Spring.Goods_Shop.Member;
+package Spring.Goods_Shop.member;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
