+++ src/main/java/Spring/Goods_Shop/member/MemberController.java
@@ -1,4 +1,4 @@
-package Spring.Goods_Shop.Member;
+package Spring.Goods_Shop.member;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
