+++ src/main/java/Spring/Goods_Shop/member/OAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package Spring.Goods_Shop.Member;
+package Spring.Goods_Shop.member;
 
 
 public interface OAuth2UserInfo {
